// src/api-client/model/pMSDataDTO.ts
/*
 * Copyright 2021 Bundesrepublik Deutschland
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Hotel Controller
 * SSI Hotel Controller
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HotelDTO } from './hotelDTO';
import { BookingDataDTO } from './bookingDataDTO';
import { CheckInCredentialDTO } from './checkInCredentialDTO';


export interface PMSDataDTO { 
    bookingData?: BookingDataDTO;
    checkInCredential?: CheckInCredentialDTO;
    hotel?: HotelDTO;
}